+++ libs/docs/core/popover/examples/popover-complex-example/popover-complex-example.component.ts
@@ -1,7 +1,14 @@
 import { Component } from '@angular/core';
 import { AvatarComponent } from '@fundamental-ngx/core/avatar';
-import { BarModule } from '@fundamental-ngx/core/bar';
-import { ButtonModule } from '@fundamental-ngx/core/button';
+import {
+    BarComponent,
+    BarElementDirective,
+    BarLeftDirective,
+    BarMiddleDirective,
+    BarRightDirective,
+    ButtonBarComponent
+} from '@fundamental-ngx/core/bar';
+import { ButtonComponent } from '@fundamental-ngx/core/button';
 import { ContentDensityDirective } from '@fundamental-ngx/core/content-density';
 import {
     PopoverBodyComponent,
@@ -10,6 +17,7 @@ import {
     PopoverComponent,
     PopoverControlComponent
 } from '@fundamental-ngx/core/popover';
+import { TitleComponent } from '@fundamental-ngx/core/title';
 
 @Component({
     selector: 'fd-popover-complex-example',
@@ -19,11 +27,17 @@ import {
     imports: [
         PopoverComponent,
         PopoverControlComponent,
-        ButtonModule,
+        BarComponent,
+        ButtonBarComponent,
+        BarElementDirective,
+        BarLeftDirective,
+        BarMiddleDirective,
+        BarRightDirective,
+        ButtonComponent,
         PopoverBodyComponent,
+        TitleComponent,
         AvatarComponent,
         PopoverBodyHeaderDirective,
-        BarModule,
         ContentDensityDirective,
         PopoverBodyFooterDirective
     ]