+++ libs/i18n/src/lib/models/lang.ts
@@ -242,6 +242,7 @@ export interface FdLanguage {
     };
     coreTokenizer: {
         moreLabel: FdLanguageKey;
+        tokenizerLabel: FdLanguageKey;
     };
     coreUploadCollection: {
         menuOkText: FdLanguageKey;